+++ examples/bmp280_example.cpp
@@ -1,4 +1,4 @@
-// BME280_example.cpp
+// BMP280_example.cpp
 
 #include <Arduino.h>
 
@@ -8,7 +8,7 @@
 
 #include "sensesp_app.h"
 #include "signalk/signalk_output.h"
-#include "sensors/bme280.h"
+#include "sensors/bmp280.h"
 
 ReactESP app([] () {
   #ifndef SERIAL_DEBUG_DISABLED
@@ -26,38 +26,31 @@ ReactESP app([] () {
   // The default is allStdSensors.
   sensesp_app = new SensESPApp(uptimeOnly);
 
-  // Create a BME280, which represents the physical sensor.
+  // Create a BMP280, which represents the physical sensor.
   // 0x77 is the default address. Some chips use 0x76, which is shown here.
-  auto* pBME280 = new BME280(0x76);
+  auto* pBMP280 = new BMP280(0x76);
 
-  // If you want to change any of the settings that are set by Adafruit_BME280::setSampling(), do
+  // If you want to change any of the settings that are set by Adafruit_BMP280::setSampling(), do
   // that here, like this:
-  // pBME280->pAdafruitBME280->setSampling(); // pass in the parameters you want
-
+  // pBMP280->pAdafruitBMP280->setSampling(); // pass in the parameters you want
 
   // Define the read_delays you're going to use:
   const uint read_delay = 1000; // once per second
   const uint pressure_read_delay = 60000; // once per minute
 
-  // Create a BME280value, which is used to read a specific value from the BME280, and send its output
+  // Create a BMP280value, which is used to read a specific value from the BMP280, and send its output
   // to SignalK as a number (float). This one is for the temperature reading.
-  auto* pBMEtemperature = new BME280value(pBME280, temperature, read_delay, "/Outside/Temperature");
+  auto* pBMPtemperature = new BMP280value(pBMP280, temperature, read_delay, "/Outside/Temperature");
       
-      pBMEtemperature->connectTo(new SKOutputNumber("environment.outside.temperature"));
+      pBMPtemperature->connectTo(new SKOutputNumber("environment.outside.temperature"));
 
 
   // Do the same for the barometric pressure value. Its read_delay is longer, since barometric pressure can't
   // change all that quickly. It could be much longer for that reason.
-  auto* pBMEpressure = new BME280value(pBME280, pressure,  pressure_read_delay, "/Outside/Pressure");
+  auto* pBMPpressure = new BMP280value(pBMP280, pressure,  pressure_read_delay, "/Outside/Pressure");
       
-      pBMEpressure->connectTo(new SKOutputNumber("environment.outside.pressure"));
+      pBMPpressure->connectTo(new SKOutputNumber("environment.outside.pressure"));
 
 
-  // Do the same for the humidity value.
-  auto* pBMEhumidity = new BME280value(pBME280, humidity,  read_delay, "Outside/Humidity");
-      
-      pBMEhumidity->connectTo(new SKOutputNumber("environment.outside.humidity"));      
-
-
-  sensesp_app->enable();
+    sensesp_app->enable();
 });