+++ CHANGELOG.md
@@ -5,6 +5,13 @@ PR's since Version 0.4.3, since most of the work done prior to that was quite lo
 to most end users.
 
 ## Pull Requests Merged Since v0.4.3
+- [#128](https://github.com/SignalK/SensESP/pull/128) (2020/05/13) feature: display IP address of ESP once connected to wifi
+- [#127](https://github.com/SignalK/SensESP/pull/127) (2020/05/13) fix: small fixes to ESP32 support; update all examples for ESP32
+- [#118](https://github.com/SignalK/SensESP/pull/118) (2020/05/25) feature: add support for ESP32
+- [#125](https://github.com/SignalK/SensESP/pull/125) (2020/05/13) fix: DigitalInputValue was not getting to SK Server
+- [#123](https://github.com/SignalK/SensESP/pull/123) (2020/05/10) feature: add SKValueListener, Threshold transform, DigitalOut sensor
+- [#117](https://github.com/SignalK/SensESP/pull/117) (2020/04/30) fix: add Wire library to library.json
+- [#113](https://github.com/SignalK/SensESP/pull/113) (2020/04/30) fix: add libraries for recent new sensors to library.json
 - [#116](https://github.com/SignalK/SensESP/pull/116) (2020/04/28) fix: correct a comment in ina219.cpp
 - [#112](https://github.com/SignalK/SensESP/pull/112) (2020/04/28) documentation: Clarify something in readme.md 
 - [#110](https://github.com/SignalK/SensESP/pull/110) (2020/04/28) fix: add Adafruit INA219 library to platformio.ini